+++ domain/src/Controller/DomainControllerBase.php
@@ -6,6 +6,7 @@ use Drupal\Core\Controller\ControllerBase;
 use Drupal\Core\Entity\EntityStorageInterface;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
+use Drupal\domain\DomainStorageInterface;
 
 /**
  * Sets a base class for injecting domain information into controllers.
@@ -21,38 +22,37 @@ class DomainControllerBase extends ControllerBase {
   /**
    * The entity storage.
    *
-   * @var \Drupal\Core\Config\Entity\ConfigEntityStorage
+   * @var \Drupal\domain\DomainStorageInterface
    */
-  protected $entityStorage;
+  protected $domainStorage;
 
   /**
    * The entity manager.
    *
    * @var \Drupal\Core\Entity\EntityTypeManagerInterface
    */
-  protected $entityManager;
+  protected $entityTypeManager;
 
   /**
    * Constructs a new DomainControllerBase.
    *
-   * @param \Drupal\Core\Entity\EntityStorageInterface $entity_storage
+   * @param \Drupal\domain\DomainStorageInterface $domain_storage
    *   The storage controller.
-   * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entity_manager
+   * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entity_type_manager
    *   The entity manager.
    */
-  public function __construct(EntityStorageInterface $entity_storage, EntityTypeManagerInterface $entity_manager) {
-    $this->entityStorage = $entity_storage;
-    $this->entityManager = $entity_manager;
+  public function __construct(DomainStorageInterface $domain_storage, EntityTypeManagerInterface $entity_type_manager) {
+    $this->domainStorage = $domain_storage;
+    $this->entityTypeManager = $entity_type_manager;
   }
 
   /**
    * {@inheritdoc}
    */
   public static function create(ContainerInterface $container) {
-    $entity_manager = $container->get('entity_type.manager');
     return new static(
-      $entity_manager->getStorage('domain'),
-      $entity_manager
+      $container->get('entity_type.manager')->getStorage('domain'),
+      $container->get('entity_type.manager')
     );
   }