+++ domain/src/DomainInterface.php
@@ -188,7 +188,7 @@ interface DomainInterface extends ConfigEntityInterface {
    * Gets the type of record match returned by the negotiator.
    *
    * This value will be set by the domain negotiation routine and is not present
-   * when loading a domain record via DomainLoaderInterface.
+   * when loading a domain record via DomainStorageInterface.
    *
    * @return int
    *   The domain record match type.