+++ domain_access/src/Tests/DomainAccessAllAffiliatesTest.php
@@ -74,7 +74,7 @@ class DomainAccessAllAffiliatesTest extends DomainTestBase {
     $this->assertText($label, 'Found the domain field instance.');
 
     // We expect to find 5 domain options.
-    $domains = \Drupal::service('domain.loader')->loadMultiple();
+    $domains = \Drupal::service('entity_type.manager')->getStorage('domain')->loadMultiple();
     foreach ($domains as $domain) {
       $string = 'value="' . $domain->id() . '"';
       $this->assertRaw($string, new FormattableMarkup('Found the %domain option.', array('%domain' => $domain->label())));